Adam's Changes are implemented

+++ lib/screens/HomeScreen.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from "react";
-import { Linking, View, Image, Text, StyleSheet, FlatList, 
-  ScrollView, TouchableOpacity, Animated } from "react-native";
+import { Alert, Linking, View, Image, Text, StyleSheet, FlatList,
+   ScrollView, TouchableOpacity, Animated } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import { StackNavigationProp } from "@react-navigation/stack";
 import { RootStackParamList } from "../../types";
@@ -10,9 +10,9 @@ interface Note {
   id: string;
   title: string;
   text: string;
+  time: string;
 }
 const user = User.getInstance();
-// console.log("User id: ", user.getId());
 
 export type HomeScreenProps = {
   navigation: any;
@@ -23,6 +23,7 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
   const [notes, setNotes] = useState<Note[]>([]);
   const [messages, setMessages] = useState<any[]>([]);
   const [isOpen, setIsOpen] = useState(true);
+  const [updateCounter, setUpdateCounter] = useState(0);
   const [drawerAnimation] = useState(new Animated.Value(0));
   const [buttonAnimation] = useState(new Animated.Value(0));
 
@@ -89,7 +90,7 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
       setNotes([...notes, route.params.note]);
     }
     fetchMessages();
-  }, [route.params]);
+  }, [route.params, updateCounter]);
 
   const fetchMessages = async () => {
     try {
@@ -109,12 +110,24 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
 
       const data = await response.json();
       setMessages(data);
-      // Map fetched messages to notes
-      const fetchedNotes: Note[] = data.map((message: any) => ({
-        id: message["@id"],
-        title: message.title || "",
-        text: message.BodyText || "", // Fallback to an empty string if 'text' is not available in the fetched message
-      }));
+
+      const fetchedNotes: Note[] = data.map((message: any) => {
+        const time = message.__rerum.isOverwritten
+          ? new Date(message.__rerum.isOverwritten)
+          : new Date(message.__rerum.createdAt);
+        time.setHours(time.getHours() - 5);
+        return {
+          id: message["@id"],
+          title: message.title || "",
+          text: message.BodyText || "",
+          time:
+            time.toLocaleString("en-US", { timeZone: "America/Chicago" }) || "",
+        };
+      });
+
+      fetchedNotes.sort(
+        (a, b) => new Date(b.time).getTime() - new Date(a.time).getTime()
+      );
       setNotes(fetchedNotes);
     } catch (error) {
       console.error("Error fetching messages:", error);
@@ -123,12 +136,14 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
 
   const addNote = (note: Note) => {
     setNotes((prevNotes) => [...prevNotes, note]);
+    setUpdateCounter(updateCounter + 1);
   };
 
   const updateNote = (note: Note) => {
     setNotes((prevNotes) =>
       prevNotes.map((prevNote) => (prevNote.id === note.id ? note : prevNote))
     );
+    setUpdateCounter(updateCounter + 1);
   };
 
   const deleteNoteFromAPI = async (id: string) => {
@@ -170,11 +185,27 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
     navigation.navigate("Login");
   };
 
-  const deleteNote = async (id: string) => {
-    const success = await deleteNoteFromAPI(id);
-    if (success) {
-      setNotes((prevNotes) => prevNotes.filter((note) => note.id !== id));
-    }
+  const deleteNote = (id: string) => {
+    Alert.alert(
+      "Delete Note",
+      "Are you sure you want to delete this note?",
+      [
+        {
+          text: "Cancel",
+          style: "cancel"
+        },
+        {
+          text: "OK",
+          onPress: async () => {
+            const success = await deleteNoteFromAPI(id);
+            if (success) {
+              setNotes((prevNotes) => prevNotes.filter((note) => note.id !== id));
+            }
+          }
+        }
+      ],
+      { cancelable: false }
+    );
   };
 
   const renderItem = ({ item }: { item: Note }) => {
@@ -185,7 +216,10 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
           navigation.navigate("EditNote", { note: item, onSave: updateNote })
         }
       >
-        <Text style={styles.noteText}>{item.title}</Text>
+        <View style={styles.noteBox}>
+          <Text style={styles.mediumText}>{item.title}</Text>
+          <Text style={styles.noteText}>{item.time}</Text>
+        </View>
         <TouchableOpacity onPress={() => deleteNote(item.id)}>
           <Ionicons name="trash-outline" size={24} color="#111111" />
         </TouchableOpacity>
@@ -195,8 +229,10 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
 
   return (
     <View style={styles.container}>
-      <TouchableOpacity onPress={toggleDrawer}
-        style={[styles.overlay, { display: !isOpen ? "flex" : "none" }]}/>
+      <TouchableOpacity
+        onPress={toggleDrawer}
+        style={[styles.overlay, { display: !isOpen ? "flex" : "none" }]}
+      />
       <Animated.View style={[styles.drawer, animatedStyles]}>
         <Animated.View style={[buttonAnimatedStyles]}>
           <TouchableOpacity style={styles.backButton} onPress={toggleDrawer}>
@@ -232,18 +268,21 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
           <Ionicons name={"log-out-outline"} size={30} color="white" />
         </TouchableOpacity>
       </Animated.View>
-      <TouchableOpacity onPress={toggleDrawer} style={styles.menuButton}>
-        <Ionicons name="menu-outline" size={24} color="white" />
-      </TouchableOpacity>
-      <View style={{ flexDirection: "row", alignItems: "center" }}>
-        <Image
-          style={styles.pfp}
-          source={require("../components/public/nopfp.png")}
-        />
-        <Text style={styles.mediumText}>Hi, {user.getName()}</Text>
+      <View style={styles.topView} >
+        <View style={{ flexDirection: "row", alignItems: "center" }} >
+          <Image
+            style={[styles.pfp, { borderRadius: 50 }]}
+            source={require("../components/public/nopfp.png")}
+          />
+          <Text style={styles.mediumText}>Hi, {user.getName()}</Text>
+        </View>
+
+        <TouchableOpacity onPress={toggleDrawer} style={[styles.menuButton]}>
+          <Ionicons name="menu-outline" size={24} color="white" />
+        </TouchableOpacity>
       </View>
 
-      <Text style={styles.title}>Field{"\n"}Notes</Text>
+      <Text style={styles.title}>Field Notes</Text>
       <ScrollView
         style={styles.filtersContainer}
         horizontal={true}
@@ -254,6 +293,9 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
           <Text style={styles.selectedFont}>All ({notes.length})</Text>
         </View>
         <View style={styles.filters}>
+          <Text style={styles.filterFont}>Most Recent</Text>
+        </View>
+        <View style={styles.filters}>
           <Text style={styles.filterFont}>Nearest</Text>
         </View>
         <View style={styles.filters}>
@@ -262,14 +304,12 @@ const HomeScreen: React.FC<HomeScreenProps> = ({ navigation, route }) => {
         <View style={styles.filters}>
           <Text style={styles.filterFont}>Alphabetical</Text>
         </View>
-        <View style={styles.filters}>
-          <Text style={styles.filterFont}>Most Recent</Text>
-        </View>
       </ScrollView>
       <FlatList
         data={notes}
         keyExtractor={(item) => item.id.toString()}
         renderItem={renderItem}
+        contentContainerStyle={{ paddingBottom: 80 }}
         ListEmptyComponent={() => (
           <View style={styles.emptyContainer}>
             <Text style={styles.emptyText}>No notes yet!</Text>
@@ -298,7 +338,7 @@ const styles = StyleSheet.create({
     height: "150%",
     flex: 1,
     zIndex: 80,
-    backgroundColor: "rgba(0, 0, 0, 0.5)", // semi-transparent black color
+    backgroundColor: "rgba(0, 0, 0, 0.5)",
   },
   drawer: {
     paddingTop: "30%",
@@ -349,8 +389,11 @@ const styles = StyleSheet.create({
     fontWeight: "600",
     maxWidth: "100%",
   },
+  noteBox: {
+    width: "100%",
+  },
   noteText: {
-    flex: 1,
+    marginLeft: "10%",
     fontSize: 18,
   },
   emptyContainer: {
@@ -374,10 +417,13 @@ const styles = StyleSheet.create({
     alignItems: "center",
     justifyContent: "center",
   },
+  topView: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    paddingHorizontal: 5,
+  },
   menuButton: {
-    position: "absolute",
-    top: 50,
-    right: 20,
     backgroundColor: "#111111",
     borderRadius: 50,
     width: 50,
@@ -393,7 +439,7 @@ const styles = StyleSheet.create({
     backgroundColor: "#fff",
     borderRadius: 20,
     marginBottom: 10,
-    maxWidth: 355,
+    width: '95%',
     padding: 20,
     paddingHorizontal: 35,
     flexDirection: "row",
@@ -402,7 +448,7 @@ const styles = StyleSheet.create({
     minHeight: 30,
     alignSelf: "center",
     borderRadius: 20,
-    paddingHorizontal: 20,
+    paddingHorizontal: 5,
     maxHeight: 30,
     marginBottom: 17,
     zIndex: 10,
@@ -435,11 +481,12 @@ const styles = StyleSheet.create({
     color: "#111111",
   },
   title: {
-    fontSize: 72,
+    fontSize: 60,
     fontWeight: "bold",
     lineHeight: 80,
     color: "#111111",
-    marginBottom: 10,
+    marginLeft: 5,
+    marginBottom: '-1%',
   },
   pfp: {
     width: 50,
